refactor : Footer 공용 컴포넌트로 수정
